Refactor logic and clean up main file
